add possibility to fetch vocabs hostes with GH-Pages
This is just a quick test to be able to pull vocabs from GitHub Pages.

Since GH-Pages does not do any Content Negotiation, I quickly (and dirty) implemented a way to pull those value lists.
Basically it checks the url and appends an /index.json or /index.index to the url.

If that fails it just uses the normal `schemaLocation`-url.

this can definitley be improved.